Added support for special instructions, and implemented JSR with unit tests. Added support for negatif memory offsets

+++ dcpu-lispasm.h
@@ -81,7 +81,7 @@ public:
     static bool is_newline(char c) { return c == '\n'; }
 
     static vector<Token> Tokenize(std::basic_istream<char>& inputStream);
-    static bool ParseOpCodeFromSexp(const SExp::Val& val, OpCode& outOpcode, uint8_t& outSpecialOp);
+    static bool ParseOpCodeFromSexp(const SExp::Val& val, OpCode& outOpcode, uint16_t& outSpecialOp);
     static void ParseValueFromSexp(const SExp::Val& val, bool isA, Value& out, uint16_t& outWord, vector<LabelRef>& foundLabels);
     static vector<Instruction> ParseTokens(const vector<Token>& tokens);
     static vector<Instruction> ParseLispAsm(char* filename);
